Add configuration for renaming rule of workspace
Basically how the naming window is decided. We can either
always use the related window (e.g. window which was moved) or
the first window in the workspace.

The related rule kinda makes sense in some situations (for example
when moving the window to a new workspace, but it also triggers
everytime when window is moved inside the workspace).

The first
window rule seems a bit more stable in this respect but we will see
which one (or something completely different rule) makes most sense.