Rename workspace when window is moved out of it
And remove the LATEST naming scheme for now. Realized that
it does not really make sense most of the time. Perhaps better
scheme would be FOCUSED which renames the workspace always
based on focused window.

The FIRST window scheme seems to make most sense except if you
move a window to an existing workspace and expect it to rename
that workspace. Not sure if this is reasonable expectation though.