Fix bug where workspace was not renamed correctly in some cases
The renaming of a workspace did not work as expected. For some reason
the find_named returned unrelated windows. In addition to this the
depth first recursion was broken and explored only first branch of the
tree...

By using window id in the search (which is available in WorkspaceReply's
ipc_data) it seems to work fine.